Created GridComponent
- gridcomponent contains all things related to gridster
