+++ src/app/components/grid/grid.component.scss
@@ -1,0 +1,3 @@
+.grid-container {
+    width: auto;
+  }
