+++ src/app/components/home/home.component.html
@@ -1,19 +1,4 @@
 <div id="wrapper">
   <app-navbar></app-navbar>
-  <gridster [options]="options" class="grid-container">
-      <gridster-item [item]="item" *ngFor="let item of dashboard">
-          <div class="gridster-item-content">
-              UnDragable text
-          </div>
-          <div class="item-buttons">
-            <x-button md-icon-button md-raised-button class="remove-button"(click)="removeItem($event, item)"
-            (touchstart)="removeItem($event, item)" mdTooltip="Remove">
-              <x-box>
-                <x-icon name="delete"></x-icon>
-                <x-label>Button</x-label>
-              </x-box>
-            </x-button>
-          </div>
-      </gridster-item>
-  </gridster>
+  <app-grid></app-grid>
 </div>
