+++ src/app/components/home/home.component.scss
@@ -6,6 +6,3 @@
   right: 0;
 }
 
-.grid-container {
-  width: auto;
-}
